fix(ops): make the MTP verify range (T<=4) bit-parity with the T=1 decode route
Root-cause the MTP losslessness violation (greedy streams diverged across
--draft-tokens k=0/1/2/3). Three defects, all in the shared T=2..4 verify
widths; the T=1 decode route is the reference arithmetic (replayssm-gdn 4.5):

- gdn conv history advanced with raw FP32 instead of the BF16-represented
  column: the multi-token conv snapshot then fed the recurrent core a state
  that the one-token path can never produce. Round the second-to-last column
  to BF16 before the in-kernel history update (T=1 publish path unchanged).
- T=4 verify jumped to the W4A4 activation-quantized routes for attn_input
  and the GDN fused plan while T=1..3 use the A16 small-t/GEMV family. Move
  the W4A4 boundaries from T>=4 to T>=5 (linear dispatch, attn_input plan,
  GDN snapshot plan) so every verify width consumes A16 inputs.
- the T>=2 NVFP4 small-t kernels accumulated one FMA chain per element while
  the T=1 GEMV runs 4 interleaved chains with a (pair*2)&3 routing; the
  different FP32 association flips rare last-ULP totals into BF16
  differences. Give the T<=4 small-t instantiations the same 4-chain layout
  (linear, GDN input, residual, attn_input, swiglu schedules); T>=5
  prefill arithmetic is unchanged.

tests/ops/test_gdn_decode_record_parity drives both production chains
(one-step snapshots vs one record plus per-commit folds) at NVFP4 27B
geometry for W=2..4 and now passes bit-exact; the GDN finite-precision
clone holds for every verify width. End to end (AIME fixture, greedy,
temp 0): k=2 and k=3 streams are now identical; k=0/k=1 still differ,
locating the remaining per-column T-dependence in the attention small-t
kernels (TokenTile-dependent CTA warp shape and per-instantiation
association). MTP acceptance rises from 27.1% to 30.2% at k=3 because the
verify target argmax is no longer perturbed by arithmetic; M1 tg128 k=3
114.35 tok/s (+2.3% vs 111.79), k=2 116.33 tok/s (+4.1%).

+++ src/ops/gdn_input_proj/gdn_conv.cuh
@@ -113,7 +113,12 @@ struct GdnConvEpilogue {
             publish.publish(token, batch_row, row, s1, s2, p);
             s0 = s1;
             s1 = s2;
-            s2 = p;
+            // The persistent convolution history is BF16: the sequential width-1 path reloads
+            // the represented value of each prior column from state, and the snapshot/record
+            // publish stores it BF16-rounded. The multi-token in-kernel history must consume the
+            // same represented value or the folded committed state is no longer a bitwise clone
+            // of the sequential one (ReplaySSM finite-precision clone contract).
+            s2 = __bfloat162float(__float2bfloat16_rn(p));
         }
     }
 };
